+++ packages/jest-resolve/src/__tests__/resolve.test.js
@@ -15,6 +15,7 @@ const path = require('path');
 const ModuleMap = require('jest-haste-map').ModuleMap;
 const Resolver = require('../');
 const userResolver = require('../__mocks__/userResolver');
+const nodeModulesPaths = require('../node_modules_paths').default;
 
 beforeEach(() => {
   userResolver.mockClear();
@@ -179,5 +180,13 @@ describe('getMockModule', () => {
       'basedir',
       path.dirname(src),
     );
+  });
+});
+
+describe('nodeModulesPaths', () => {
+  it('provides custom module paths after node_modules', () => {
+    const src = require.resolve('../');
+    const result = nodeModulesPaths(src, {paths: ['./customFolder']});
+    expect(result[result.length - 1]).toBe('./customFolder');
   });
 });